editor2 is no more
